+++ Tests/SentryTests/Integrations/SentrySessionGeneratorTests.swift
@@ -11,6 +11,7 @@ class SentrySessionGeneratorTests: XCTestCase {
         var healthy = 0
         var errored = 0
         var crashed = 0
+        var oom = 0
         var abnormal = 0
     }
     
@@ -30,17 +31,29 @@ class SentrySessionGeneratorTests: XCTestCase {
             fileManager = try SentryFileManager(dsn: dsn, andCurrentDateProvider: TestCurrentDateProvider())
             
             fileManager.deleteCurrentSession()
+            fileManager.deleteCrashedSession()
             fileManager.deleteTimestampLastInForeground()
+            fileManager.deleteAppState()
         } catch {
             XCTFail("Could not delete session data")
         }
+    }
+    
+    override func tearDown() {
+        super.tearDown()
+        
+        fileManager.deleteCurrentSession()
+        fileManager.deleteCrashedSession()
+        fileManager.deleteTimestampLastInForeground()
+        fileManager.deleteAppState()
+        autoSessionTrackingIntegration.stop()
     }
     
     /**
      * Disabled on purpose. This test just sends sessions to Sentry, but doesn't verify that they arrive there properly.
      */
     func tesSendSessions() {
-        sendSessions(amount: Sessions(healthy: 10, errored: 10, crashed: 3, abnormal: 1))
+        sendSessions(amount: Sessions(healthy: 10, errored: 10, crashed: 3, oom: 1, abnormal: 1))
     }
     
     private func sendSessions(amount: Sessions ) {
@@ -85,6 +98,24 @@ class SentrySessionGeneratorTests: XCTestCase {
         }
         sentryCrash.internalCrashedLastLaunch = false
         
+        #if os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
+        let appState = SentryAppState(releaseName: options.releaseName!, osVersion: UIDevice.current.systemVersion, isDebugging: false)
+        appState.isActive = true
+        fileManager.store(appState)
+        
+        for _ in Array(1...amount.oom) {
+            // send crashed session
+            crashIntegration.install(with: options)
+            
+            autoSessionTrackingIntegration.stop()
+            autoSessionTrackingIntegration.install(with: options)
+            goToForeground()
+            
+            SentrySDK.captureCrash(TestData.oomEvent)
+        }
+        fileManager.deleteAppState()
+        #endif
+        
         for _ in Array(1...amount.abnormal) {
             autoSessionTrackingIntegration.stop()
             autoSessionTrackingIntegration.install(with: options)
@@ -119,7 +150,7 @@ class SentrySessionGeneratorTests: XCTestCase {
         let hub = SentryHub(client: client, andScope: nil, andCrashAdapter: self.sentryCrash)
         SentrySDK.setCurrentHub(hub)
         
-        crashIntegration = SentryCrashIntegration(crashWrapper: sentryCrash, andDispatchQueueWrapper: TestSentryDispatchQueueWrapper())
+        crashIntegration = SentryCrashIntegration(crashAdapter: sentryCrash, andDispatchQueueWrapper: TestSentryDispatchQueueWrapper())
         crashIntegration.install(with: options)
         
         autoSessionTrackingIntegration = SentryAutoSessionTrackingIntegration()