+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+- fix: Crash when passing garbage to maxBreadcrumbs #1018
+- fix: OutOfMemory exception type #1015
+- fix: macOS version for Mac Catalyst #1011
+
+## 7.0.0-alpha.3
+
+- feat: Out Of Memory Tracking #1001
+
 ## 7.0.0-alpha.2
 
 ### Features